wip: splitting enterprise code out more
