+++ enterprise/frontend/src/metabase-enterprise/caching/components/StrategyFormLauncher.tsx
@@ -1,14 +1,13 @@
 import { t } from "ttag";
 import { findWhere } from "underscore";
 
+import { rootId } from "metabase/admin/performance/constants";
+import type { Config, UpdateTargetId } from "metabase/admin/performance/types";
+import { getShortStrategyLabel } from "metabase/admin/performance/utils";
 import { color } from "metabase/lib/colors";
 import { FixedSizeIcon, Flex, Title, Tooltip, useHover } from "metabase/ui";
 
-import { rootId } from "../constants";
-import type { Config, UpdateTargetId } from "../types";
-import { getShortStrategyLabel } from "../types";
-
-import { PolicyToken } from "./StrategyEditorForDatabases.styled";
+import { PolicyToken } from "./StrategyFormLauncher.styled";
 
 export const StrategyFormLauncher = ({
   forId,
